# apps/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Atarimae API",
    "version": "0.0.0",
    "description": "Self-hosted communication board for small teams.\n\nAll timestamps are UTC ISO 8601. All identifiers are UUIDv7.",
    "license": {
      "name": "AGPL-3.0-only",
      "url": "https://www.gnu.org/licenses/agpl-3.0.html"
    }
  },
...
  "paths": {
...
    "/setup/status": {
      "get": {
        "summary": "Whether the organisation has been initialised",
        "tags": [
          "setup"
        ],
        "description": "Unauthenticated. Reveals only whether an Owner exists, which the first-run screen needs in order to decide what to show.",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "initialized"
                  ],
                  "properties": {
                    "initialized": {
                      "description": "True once an Owner exists. The setup endpoint then refuses further calls.",
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "description": "Stable machine-readable error code.",
                      "type": "string",
                      "example": "NO_ELIGIBLE_RECIPIENTS"
                    },
                    "message": {
                      "description": "Developer-facing English description. Not shown to end users.",
                      "type": "string"
                    },
                    "details": {
                      "description": "Structured context. Never contains secrets.",
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "description": "Correlates with server logs and audit_logs.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "description": "Stable machine-readable error code.",
                      "type": "string",
                      "example": "NO_ELIGIBLE_RECIPIENTS"
                    },
                    "message": {
                      "description": "Developer-facing English description. Not shown to end users.",
                      "type": "string"
                    },
                    "details": {
                      "description": "Structured context. Never contains secrets.",
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "description": "Correlates with server logs and audit_logs.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "description": "Stable machine-readable error code.",
                      "type": "string",
                      "example": "NO_ELIGIBLE_RECIPIENTS"
                    },
                    "message": {
                      "description": "Developer-facing English description. Not shown to end users.",
                      "type": "string"
                    },
                    "details": {
                      "description": "Structured context. Never contains secrets.",
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "description": "Correlates with server logs and audit_logs.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "description": "Stable machine-readable error code.",
                      "type": "string",
                      "example": "NO_ELIGIBLE_RECIPIENTS"
                    },
                    "message": {
                      "description": "Developer-facing English description. Not shown to end users.",
                      "type": "string"
                    },
                    "details": {
                      "description": "Structured context. Never contains secrets.",
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "description": "Correlates with server logs and audit_logs.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "description": "Stable machine-readable error code.",
                      "type": "string",
                      "example": "NO_ELIGIBLE_RECIPIENTS"
                    },
                    "message": {
                      "description": "Developer-facing English description. Not shown to end users.",
                      "type": "string"
                    },
                    "details": {
                      "description": "Structured context. Never contains secrets.",
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "description": "Correlates with server logs and audit_logs.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "413": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "description": "Stable machine-readable error code.",
                      "type": "string",
                      "example": "NO_ELIGIBLE_RECIPIENTS"
                    },
                    "message": {
                      "description": "Developer-facing English description. Not shown to end users.",
                      "type": "string"
                    },
                    "details": {
                      "description": "Structured context. Never contains secrets.",
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "description": "Correlates with server logs and audit_logs.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "415": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "description": "Stable machine-readable error code.",
                      "type": "string",
                      "example": "NO_ELIGIBLE_RECIPIENTS"
                    },
                    "message": {
                      "description": "Developer-facing English description. Not shown to end users.",
                      "type": "string"
                    },
                    "details": {
                      "description": "Structured context. Never contains secrets.",
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "description": "Correlates with server logs and audit_logs.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "description": "Stable machine-readable error code.",
                      "type": "string",
                      "example": "NO_ELIGIBLE_RECIPIENTS"
                    },
                    "message": {
                      "description": "Developer-facing English description. Not shown to end users.",
                      "type": "string"
                    },
                    "details": {
                      "description": "Structured context. Never contains secrets.",
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "description": "Correlates with server logs and audit_logs.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "description": "Stable machine-readable error code.",
                      "type": "string",
                      "example": "NO_ELIGIBLE_RECIPIENTS"
                    },
                    "message": {
                      "description": "Developer-facing English description. Not shown to end users.",
                      "type": "string"
                    },
                    "details": {
                      "description": "Structured context. Never contains secrets.",
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "description": "Correlates with server logs and audit_logs.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "description": "Stable machine-readable error code.",
                      "type": "string",
                      "example": "NO_ELIGIBLE_RECIPIENTS"
                    },
                    "message": {
                      "description": "Developer-facing English description. Not shown to end users.",
                      "type": "string"
                    },
                    "details": {
                      "description": "Structured context. Never contains secrets.",
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "description": "Correlates with server logs and audit_logs.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Version 1"
    }
  ],
...
}